Saved recipe: TEST.compute_nyc_trips_tip_percentage

+++ recipes/compute_nyc_trips_tip_percentage.sql
@@ -1,5 +1,5 @@
 SELECT 
     *,
-    (tip_amount / total_amount) * 100 AS tip_percentage
+    (tip_amount::int / total_amount::int) * 100 AS tip_percentage
 FROM 
     "public"."SQL_nyc_trips_copy"
